divided home page into classes
