+++ lib/Screens/HomePageScreen/home_app_bar.dart
@@ -1,0 +1,47 @@
+import 'package:flutter/material.dart';
+
+const String homePage = 'Home Page';
+
+class HomeAppBar extends StatelessWidget {
+  const HomeAppBar({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return AppBar(
+      automaticallyImplyLeading: false,
+      elevation: 0,
+      title: SafeArea(
+        child: Padding(
+          padding: const EdgeInsets.only(left: 10, right: 10),
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: [
+              // ignore: prefer_const_constructors
+              Text(
+                homePage,
+                style: const TextStyle(
+                  fontSize: 20,
+                  color: Colors.white,
+                  fontWeight: FontWeight.bold,
+                ),
+              ),
+              IconButton(
+                  onPressed: () {
+                    // Navigator.push(
+                    //     context,
+                    //     PageTransition(
+                    //         alignment: Alignment.topRight,
+                    //         child: const SettingsPage(),
+                    //         type: PageTransitionType.scale),);
+                  },
+                  icon: const Icon(
+                    Icons.settings,
+                    color: Colors.white,
+                  ))
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
